requirements.txt
python-dotenv==0.9.1
requests==2.31.0
urllib3==2.2.1
